+++ Tests/BarkVisorTests/HelperProtocolConformanceTests.swift
@@ -39,11 +39,6 @@
                 func getAllBridgeStates(reply: @escaping (String) -> Void) {
                     reply("[]")
                 }
-                func installUpdate(
-                    packagePath: String, expectedVersion: String, reply: @escaping (Bool, String?) -> Void,
-                ) {
-                    reply(false, nil)
-                }
             }
 
             let handler = MinimalHandler()